Add workarounds for ifort 15.0 bug.
The bug relates to passing a C_PTR function result to a BIND(C)
procedure where the associated dummy argument has the VALUE
attribute.  The workaround in all cases has been to rename the
function result variable to a temporary (original name with _tmp
added), add the appropriate declaration for that temporary and then
assign that temporary immediately after the variable that was the
original function result is been defined.