Workarounds for an ICE with ifort 15.0
The ice is associated with a C_PTR argument being passed from a
main program through an external procedure to the gtk_widget_queue_draw
function.  The workaround is to replace the argument with a module
variable.  Once the ICE is fixed the module can be removed and the
module variable changed back to being a local variable of the program
and an argument of the external procedure.